Formulario Especie
Listos los 3 administradores y las vistas

+++ Mariposario/Mariposario/FrmGenero.cs
@@ -12,75 +12,91 @@ namespace Mariposario
 {
     public partial class FrmGenero : Form
     {
-        public List<Genero> listaGeneros;
+        public List<Genero> listaGeneros { set; get; }
+        List<Especie> listaEspecie { set; get; }
+        public frmMain FrmMain { set; get; }
         Conexion conexion = new Conexion();
 
         public FrmGenero()
         {
             InitializeComponent();
-            //borrar
-            listaGeneros = conexion.obtenerGeneros();
-            lbxGenero.Items.AddRange(listaGeneros.ToArray());
+            //listaGeneros = conexion.obtenerGeneros();
+            //lbxGenero.Items.AddRange(listaGeneros.ToArray());
         }
         #region Constructor que recibe una lista de generos
-        public FrmGenero(List<Genero> listaGeneros) {
-            listaGeneros = conexion.obtenerGeneros();
-            lbxGenero.Items.AddRange(listaGeneros.ToArray());
+        public FrmGenero(frmMain frmMain) {
+            InitializeComponent();
+            this.FrmMain = frmMain;
+            this.listaEspecie = this.FrmMain.Especies;
+            listaGeneros = FrmMain.Generos;
+            if( listaGeneros.Count!=0 )
+                lbxGenero.Items.AddRange(listaGeneros.ToArray());
         }
         #endregion
-
         #region Carga los generos y las especies que corresponden que corresponden a un genero
         private void lbxGenero_SelectedIndexChanged(object sender, EventArgs e)
         {
             this.lbxInformacion.Items.Clear();
             Genero genero = (Genero)lbxGenero.SelectedItem;
-            if(genero!=null)
+            if (genero != null)
+                this.txtCodigo.Text = genero.Identificador;
               this.lbxInformacion.Items.AddRange(genero.Especies.ToArray());
         }
         #endregion
-
+        #region AgregarGenero
         private void btnAgregar_Click(object sender, EventArgs e)
         {
-            FrmAgregar_Especie_Genero_FamiliaNatural f= new FrmAgregar_Especie_Genero_FamiliaNatural(new Genero("",""));
+            FrmAgregar_Especie_Genero_FamiliaNatural f= new FrmAgregar_Especie_Genero_FamiliaNatural(new Genero("",""),FrmMain);
             f.ShowDialog();
             if (f.DialogResult == DialogResult.OK) {
                 this.actualizaElLisboxGeneros();
             }
         }
-
+        #endregion
         #region ACTUALIZAR LISTBOX GENEROS
         private void actualizaElLisboxGeneros()
         {
-            listaGeneros = conexion.obtenerGeneros();
+            this.FrmMain.CargarBaseDeDatos();
+            listaGeneros = this.FrmMain.Generos;
             this.lbxGenero.Items.Clear();
             this.lbxGenero.Items.AddRange(listaGeneros.ToArray());
 
         }
         #endregion
-
+        #region Actualizar Género
         private void btnActualizar_Click(object sender, EventArgs e)
         {
             if ( this.lbxGenero.SelectedItem != null)
             {
-                Genero g = (Genero)lbxGenero.SelectedItem;
-                FrmActualizar f = new FrmActualizar(g);
-                f.ShowDialog();
-                if ( f.DialogResult == DialogResult.OK ) {
-                    this.actualizaElLisboxGeneros();
-                }
+                    Genero g = (Genero)lbxGenero.SelectedItem;
+                    FrmActualizar f = new FrmActualizar(g,this.FrmMain);
+                    f.ShowDialog();
+                    if (f.DialogResult == DialogResult.OK)
+                    {
+                        MessageBox.Show("Actualización Exitosa", "Actualizar género...", MessageBoxButtons.OK);
+                        this.actualizaElLisboxGeneros();
+
+                    }
             }
             else {
-                MessageBox.Show("Seleccione un elemento");
+                MessageBox.Show("Seleccione un elemento", "Género", MessageBoxButtons.OK);
             }
         }
-
+        #endregion
+        #region Eliminar Género
         private void btnEliminar_Click(object sender, EventArgs e)
         {
             if (this.lbxGenero.SelectedItem != null) {
-                Genero g = (Genero)lbxGenero.SelectedItem;
-                if (conexion.eliminarGenero(g.Identificador)) {
-                    this.actualizaElLisboxGeneros();
-                    MessageBox.Show("Eliminación Exitosa");
+
+                DialogResult resultado = MessageBox.Show("¿Está seguro de eliminar este elemento?", "Eliminar género", MessageBoxButtons.YesNoCancel);
+                if (resultado == System.Windows.Forms.DialogResult.Yes)
+                {
+                    Genero g = (Genero)lbxGenero.SelectedItem;
+                    if (conexion.eliminarGenero(g.Identificador))
+                    {
+                        this.actualizaElLisboxGeneros();
+                        MessageBox.Show("Eliminación Exitosa", "Eliminar género...", MessageBoxButtons.OK);
+                    }
                 }
                 else
                 {
@@ -93,6 +109,6 @@ namespace Mariposario
                 MessageBox.Show("Seleccione una Género");
             }
         }
-
+        #endregion
     }
 }
